Port center-of-rotation UX to camera_lidar_trajectory_viewer's OrbitCamera
raylib_widgets::OrbitCamera (shared with camera_lidar_calibration) gains an
eased target transition, ground-plane ray picking (Ctrl+Right-click or
middle-click), a shared "Center of rotation" dialog, and a 3D crosshair
marker at the current center -- porting multi_view_tls_registration's
center-of-rotation UX without replacing OrbitCamera's Camera3D-based
rendering pipeline. camera_lidar_trajectory_viewer's local Orbit struct is
replaced by the shared OrbitCamera, and its State struct is renamed to
AppState.

Also finishes/fixes the in-progress raylib_widgets refactor found along the
way: builds OrbitCamera.cpp into the library (was listed on disk but never
compiled), removes camera_lidar_calibration's now-duplicate OrbitCamera
definitions (would have been a link error once OrbitCamera.cpp actually
built), and reverts an unfinished move of multi_view_tls_registration's
rl_utils.h/.cpp into raylib_widgets -- it stays local to that app per its own
"deliberately not shared" comment, and the move had left
TrajectoryViewer.cpp with a duplicate app_state definition (compile error)
plus missing Core/Eigen include paths.
