+++ apps/camera_lidar_calibration/Renderer.cpp
@@ -3,8 +3,10 @@
 #include "raymath.h"
 // glad function pointers are compiled into raylib; the header only declares them
 #include "external/glad.h"
+#include "RendererShaders.h"
 #include <cmath>
 #include <algorithm>
+#include <string>
 #include <vector>
 
 // ── Jet colormap ─────────────────────────────────────────────────────────────
@@ -21,52 +23,10 @@ Color jetColor(float t) {
     };
 }
 
-// ── OrbitCamera ───────────────────────────────────────────────────────────────
-Camera3D OrbitCamera::toRaylib() const {
-    float az  = azimuth   * (float)DEG2RAD;
-    float el  = elevation * (float)DEG2RAD;
-    Vector3 pos = {
-        target.x + distance * std::cos(el) * std::sin(az),
-        target.y + distance * std::sin(el),
-        target.z + distance * std::cos(el) * std::cos(az)
-    };
-    Camera3D cam;
-    cam.position   = pos;
-    cam.target     = target;
-    cam.up         = {0.f, 1.f, 0.f};
-    cam.fovy       = 45.f;
-    cam.projection = CAMERA_PERSPECTIVE;
-    return cam;
-}
-
-void OrbitCamera::update(bool active) {
-    if (!active) return;
-
-    // Left-drag → orbit
-    if (IsMouseButtonDown(MOUSE_BUTTON_LEFT)) {
-        Vector2 d = GetMouseDelta();
-        azimuth   -= d.x * 0.4f;
-        elevation += d.y * 0.4f;
-        elevation  = std::max(-89.f, std::min(89.f, elevation));
-    }
-    // Right-drag → pan
-    if (IsMouseButtonDown(MOUSE_BUTTON_RIGHT)) {
-        Camera3D cam = toRaylib();
-        Vector3 fwd   = Vector3Normalize(Vector3Subtract(cam.target, cam.position));
-        Vector3 right = Vector3Normalize(Vector3CrossProduct(fwd, cam.up));
-        Vector3 up    = Vector3CrossProduct(right, fwd);
-        Vector2 d     = GetMouseDelta();
-        float   speed = distance * 0.002f;
-        target = Vector3Add(target, Vector3Scale(right, -d.x * speed));
-        target = Vector3Add(target, Vector3Scale(up,     d.y * speed));
-    }
-    // Scroll → zoom
-    float wheel = GetMouseWheelMove();
-    if (wheel != 0.f) {
-        distance -= wheel * distance * 0.1f;
-        distance  = std::max(0.5f, distance);
-    }
-}
+// OrbitCamera::toRaylib()/update() now live in raylib_widgets/src/OrbitCamera.cpp
+// (Renderer.h aliases raylib_widgets::OrbitCamera) -- this used to duplicate
+// those definitions, which became a duplicate-symbol link error once
+// raylib_widgets/CMakeLists.txt started actually compiling that .cpp.
 
 // World-frame convention: E.rx/ry/rz = camera orientation in world (R_wc, ZYX Euler).
 // E.tx/ty/tz = camera position in world.  p_cam = R_wc^T * (p_lidar - C).
@@ -105,147 +65,13 @@ void Renderer::shutdown() {
     }
 }
 
-// ── GPU point cloud shaders ──────────────────────────────────────────────────
-// Explicit attribute locations so one VAO works with both programs:
-// location 0 = position (raylib coords), location 1 = intensity.
-static const char* kPointVS = R"(
-#version 330
-layout(location = 0) in vec3 vertexPosition;
-layout(location = 1) in float vertexIntensity;
-uniform mat4 mvp;
-uniform float pointSize;
-uniform mat4 lidarToCam;   // extrinsics (for RGB mode)
-uniform vec4 K;            // fx, fy, cx, cy
-uniform vec2 imgSize;
-out vec3 fragPos;
-out float fragIntensity;
-out vec2 fragUV;
-out float fragCamDepth;
-void main() {
-    fragPos = vertexPosition;
-    fragIntensity = vertexIntensity;
-    gl_Position = mvp * vec4(vertexPosition, 1.0);
-    gl_PointSize = pointSize;
-
-    // Project into the camera image for RGB sampling (rectified → pinhole)
-    vec3 lidar = vec3(vertexPosition.x, -vertexPosition.z, vertexPosition.y);
-    vec3 pc = (lidarToCam * vec4(lidar, 1.0)).xyz;
-    fragCamDepth = pc.z;
-    vec2 uv = (K.xy * (pc.xy / max(pc.z, 1e-6)) + K.zw) / imgSize;
-    fragUV = uv;
-}
-)";
-
-static const char* kPointFS = R"(
-#version 330
-in vec3 fragPos;
-in float fragIntensity;
-in vec2 fragUV;
-in float fragCamDepth;
-uniform int colorMode;      // 0 = distance, 1 = intensity, 2 = height, 3 = camera RGB
-uniform vec2 heightRange;   // min/max of raylib Y (lidar Z)
-uniform float maxDist;
-uniform float opacity;
-uniform sampler2D imageTex;
-out vec4 finalColor;
-
-vec3 jet(float t) {
-    t = clamp(t, 0.0, 1.0);
-    return clamp(vec3(1.5 - abs(4.0*t - 3.0),
-                      1.5 - abs(4.0*t - 2.0),
-                      1.5 - abs(4.0*t - 1.0)), 0.0, 1.0);
-}
-
-void main() {
-    if (colorMode == 3) {
-        bool seen = fragCamDepth > 0.0
-            && fragUV.x >= 0.0 && fragUV.x <= 1.0
-            && fragUV.y >= 0.0 && fragUV.y <= 1.0;
-        // points the camera cannot see stay gray — shows the camera FOV
-        vec3 c = seen ? texture(imageTex, fragUV).rgb : vec3(0.25);
-        finalColor = vec4(c, opacity);
-        return;
-    }
-    float t;
-    if (colorMode == 1)
-        t = fragIntensity;
-    else if (colorMode == 2)
-        t = (fragPos.y - heightRange.x) / max(heightRange.y - heightRange.x, 1e-6);
-    else
-        t = length(fragPos) / max(maxDist, 1e-6);
-    finalColor = vec4(jet(t), opacity);
-}
-)";
-
-// Projects lidar points directly onto the image plane. Position attribute is
-// in raylib coords, converted back to lidar frame here. With w = z_cam the
-// hardware clip rejects points behind the camera; optional rational+tangential
-// distortion handles non-rectified images (pass zeros when rectified).
-static const char* kProjVS = R"(
-#version 330
-layout(location = 0) in vec3 vertexPosition;
-layout(location = 1) in float vertexIntensity;
-uniform mat4 lidarToCam;   // extrinsics
-uniform vec4 K;            // fx, fy, cx, cy
-uniform vec2 imgSize;
-uniform vec3 kRad1;        // k1 k2 k3
-uniform vec3 kRad2;        // k4 k5 k6
-uniform vec2 pTan;         // p1 p2
-uniform float pointSize;
-out float fragDepth;
-out float fragIntensity;
-void main() {
-    // raylib coords -> lidar: x = rx, y = -rz, z = ry
-    vec3 lidar = vec3(vertexPosition.x, -vertexPosition.z, vertexPosition.y);
-    vec3 pc = (lidarToCam * vec4(lidar, 1.0)).xyz;
-    fragDepth = pc.z;
-    fragIntensity = vertexIntensity;
-
-    vec2 n = pc.xy / max(pc.z, 1e-6);
-    float r2 = dot(n, n);
-    float radial = (1.0 + kRad1.x*r2 + kRad1.y*r2*r2 + kRad1.z*r2*r2*r2)
-                 / (1.0 + kRad2.x*r2 + kRad2.y*r2*r2 + kRad2.z*r2*r2*r2);
-    vec2 d = n * radial
-           + vec2(2.0*pTan.x*n.x*n.y + pTan.y*(r2 + 2.0*n.x*n.x),
-                  pTan.x*(r2 + 2.0*n.y*n.y) + 2.0*pTan.y*n.x*n.y);
-    vec2 uv = K.xy * d + K.zw;                     // pixel coords
-
-    // pixel -> clip space (y down, like raylib's render-texture ortho)
-    gl_Position = vec4((2.0*uv.x/imgSize.x - 1.0) * pc.z,
-                       -(2.0*uv.y/imgSize.y - 1.0) * pc.z,
-                       0.0,
-                       pc.z);
-    gl_PointSize = pointSize;
-}
-)";
-
-static const char* kProjFS = R"(
-#version 330
-in float fragDepth;
-in float fragIntensity;
-uniform vec2 depthRange;
-uniform float opacity;
-uniform int colorMode;
-out vec4 finalColor;
-
-vec3 jet(float t) {
-    t = clamp(t, 0.0, 1.0);
-    return clamp(vec3(1.5 - abs(4.0*t - 3.0),
-                      1.5 - abs(4.0*t - 2.0),
-                      1.5 - abs(4.0*t - 1.0)), 0.0, 1.0);
-}
-
-void main() {
-    if (fragDepth < depthRange.x || fragDepth > depthRange.y) discard;
-    float t = (colorMode == 1)
-        ? fragIntensity
-        : (fragDepth - depthRange.x) / max(depthRange.y - depthRange.x, 1e-6);
-    finalColor = vec4(jet(t), opacity);
-}
-)";
+using renderer_shaders::kPointVS;
+using renderer_shaders::kPointFS;
+using renderer_shaders::kProjVS;
+using renderer_shaders::kProjFS;
 
 void Renderer::initPointShader() {
-    pointShader = LoadShaderFromMemory(kPointVS, kPointFS);
+    pointShader = LoadShaderFromMemory(kPointVS, kPointFS.c_str());
     shaderValid = pointShader.id > 0;
     if (!shaderValid) {
         TraceLog(LOG_ERROR, "Point cloud shader failed to compile");
@@ -262,7 +88,7 @@ void Renderer::initPointShader() {
         locCamTex      = rlGetLocationUniform(pointShader.id, "imageTex");
     }
 
-    projShader = LoadShaderFromMemory(kProjVS, kProjFS);
+    projShader = LoadShaderFromMemory(kProjVS, kProjFS.c_str());
     projShaderValid = projShader.id > 0;
     if (!projShaderValid) {
         TraceLog(LOG_ERROR, "Projection shader failed to compile");
