+++ core/src/raylib_render.cpp
@@ -5,10 +5,13 @@
 #include "raymath.h"
 #include "rlgl.h"
 
+#include "raylib_render_shaders.hpp"
+
 #include <algorithm>
 #include <cmath>
 #include <cstdio>
 #include <limits>
+#include <string>
 
 namespace
 {
@@ -25,74 +28,8 @@ namespace
 // ============================================================================
 namespace
 {
-    // vertexIntensity is per-point LAS/LAZ intensity, normalized to [0,1] per
-    // scan at upload time (see ScanRenderer::rebuild). vertexPosition is
-    // already world-space (rebuild() applies pc.m_pose before uploading), so
-    // it doubles as the Elevation/Distance color modes' input with no extra
-    // per-vertex data needed. colorMode selects between the flat per-scan
-    // pointColor (0, the original app's only mode) and the ScanColorMode
-    // gradients (1/2/3), matching the enum's Intensity/Elevation/Distance
-    // ordering exactly (see ScanRenderer::draw()'s static_cast below).
-    const char* kPointVS = R"(
-#version 330
-in vec3 vertexPosition;
-in float vertexIntensity;
-uniform mat4 mvp;
-uniform float pointSize;
-out float fragIntensity;
-out vec3 fragWorldPos;
-void main()
-{
-    gl_Position = mvp * vec4(vertexPosition, 1.0);
-    gl_PointSize = pointSize;
-    fragIntensity = vertexIntensity;
-    fragWorldPos = vertexPosition;
-}
-)";
-
-    const char* kPointFS = R"(
-#version 330
-uniform vec4 pointColor;
-uniform int colorMode;
-uniform float elevMin;
-uniform float elevMax;
-uniform vec3 distCenter;
-uniform float distMax;
-in float fragIntensity;
-in vec3 fragWorldPos;
-out vec4 finalColor;
-
-vec3 jet(float t)
-{
-    t = clamp(t, 0.0, 1.0);
-    float r = clamp(1.5 - abs(4.0 * t - 3.0), 0.0, 1.0);
-    float g = clamp(1.5 - abs(4.0 * t - 2.0), 0.0, 1.0);
-    float b = clamp(1.5 - abs(4.0 * t - 1.0), 0.0, 1.0);
-    return vec3(r, g, b);
-}
-
-void main()
-{
-    if (colorMode == 1)
-    {
-        finalColor = vec4(jet(fragIntensity), pointColor.a);
-    }
-    else if (colorMode == 2)
-    {
-        float range = max(elevMax - elevMin, 1e-6);
-        finalColor = vec4(jet((fragWorldPos.z - elevMin) / range), pointColor.a);
-    }
-    else if (colorMode == 3)
-    {
-        float d = length(fragWorldPos - distCenter);
-        finalColor = vec4(jet(d / max(distMax, 1e-6)), pointColor.a);
-    }
-    else
-    {
-        finalColor = pointColor;
-    }
-}
-)";
+    using raylib_render_shaders::kPointVS;
+    using raylib_render_shaders::kPointFS;
 
     // Bytes per vertex in the uploaded buffer (xyz position + intensity).
     constexpr int kVertexStride = 4 * sizeof(float);
@@ -105,7 +42,7 @@ ScanRenderer::~ScanRenderer()
 
 void ScanRenderer::init()
 {
-    shader_ = LoadShaderFromMemory(kPointVS, kPointFS);
+    shader_ = LoadShaderFromMemory(kPointVS, kPointFS.c_str());
     shaderValid_ = shader_.id > 0;
     if (shaderValid_)
     {
